talkative_phone/pyproject.toml
[tool.poetry]
name = "talkative_phone"
version = "0.1.0"
description = ""
authors = ["MaxST <mstolpasov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.1"
sqlalchemy = "^1.3"
sqlalchemy-utils = "^0.34.1"
passlib = "^1.7"
pyqt5 = "^5.13"
dynaconf = "^2.0"
tabulate = "^0.8.3"
pycryptodomex = "^3.8"
Pillow = "^6.1"
pymongo = "^3.9"
mongoengine = "^0.18.2"
python-dateutil = "^2.8"
docutils = "^0.15.2"
pygments = "^2.4"
kivy = "^1.11"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.0"
pytest = "^5.0"
pytest-cov = "^2.7"
mccabe = "^0.6.1"
Sphinx = "^2.1"
sphinx_rtd_theme = "^0.4.3"
sphinx-issues = "^1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
